+++ src/hsm/cvc.h
@@ -28,78 +28,78 @@
 
 typedef struct PUK {
     const uint8_t *puk;
-    size_t puk_len;
+    uint16_t puk_len;
     const uint8_t *car;
-    size_t car_len;
+    uint16_t car_len;
     const uint8_t *chr;
-    size_t chr_len;
+    uint16_t chr_len;
     const uint8_t *cvcert;
-    size_t cvcert_len;
+    uint16_t cvcert_len;
     bool copied;
 } PUK;
 
 #define MAX_PUK_STORE_ENTRIES 4
 
-extern size_t asn1_cvc_cert(void *rsa_ecdsa,
+extern uint16_t asn1_cvc_cert(void *rsa_ecdsa,
                             uint8_t key_type,
                             uint8_t *buf,
-                            size_t buf_len,
+                            uint16_t buf_len,
                             const uint8_t *ext,
-                            size_t ext_len,
+                            uint16_t ext_len,
                             bool full);
-extern size_t asn1_cvc_aut(void *rsa_ecdsa,
+extern uint16_t asn1_cvc_aut(void *rsa_ecdsa,
                            uint8_t key_type,
                            uint8_t *buf,
-                           size_t buf_len,
+                           uint16_t buf_len,
                            const uint8_t *ext,
-                           size_t ext_len);
-extern size_t asn1_build_cert_description(const uint8_t *label,
-                                          size_t label_len,
+                           uint16_t ext_len);
+extern uint16_t asn1_build_cert_description(const uint8_t *label,
+                                          uint16_t label_len,
                                           const uint8_t *puk,
-                                          size_t puk_len,
+                                          uint16_t puk_len,
                                           uint16_t fid,
                                           uint8_t *buf,
-                                          size_t buf_len);
-extern const uint8_t *cvc_get_field(const uint8_t *data, size_t len, size_t *olen, uint16_t tag);
-extern const uint8_t *cvc_get_car(const uint8_t *data, size_t len, size_t *olen);
-extern const uint8_t *cvc_get_chr(const uint8_t *data, size_t len, size_t *olen);
-extern const uint8_t *cvc_get_pub(const uint8_t *data, size_t len, size_t *olen);
-extern const uint8_t *cvc_get_ext(const uint8_t *data, size_t len, size_t *olen);
-extern int cvc_verify(const uint8_t *cert, size_t cert_len, const uint8_t *ca, size_t ca_len);
-extern mbedtls_ecp_group_id cvc_inherite_ec_group(const uint8_t *ca, size_t ca_len);
+                                          uint16_t buf_len);
+extern const uint8_t *cvc_get_field(const uint8_t *data, uint16_t len, uint16_t *olen, uint16_t tag);
+extern const uint8_t *cvc_get_car(const uint8_t *data, uint16_t len, uint16_t *olen);
+extern const uint8_t *cvc_get_chr(const uint8_t *data, uint16_t len, uint16_t *olen);
+extern const uint8_t *cvc_get_pub(const uint8_t *data, uint16_t len, uint16_t *olen);
+extern const uint8_t *cvc_get_ext(const uint8_t *data, uint16_t len, uint16_t *olen);
+extern int cvc_verify(const uint8_t *cert, uint16_t cert_len, const uint8_t *ca, uint16_t ca_len);
+extern mbedtls_ecp_group_id cvc_inherite_ec_group(const uint8_t *ca, uint16_t ca_len);
 extern int puk_verify(const uint8_t *sig,
-                      size_t sig_len,
+                      uint16_t sig_len,
                       const uint8_t *hash,
-                      size_t hash_len,
+                      uint16_t hash_len,
                       const uint8_t *ca,
-                      size_t ca_len);
-extern size_t asn1_build_prkd_ecc(const uint8_t *label,
-                                  size_t label_len,
+                      uint16_t ca_len);
+extern uint16_t asn1_build_prkd_ecc(const uint8_t *label,
+                                  uint16_t label_len,
                                   const uint8_t *keyid,
-                                  size_t keyid_len,
-                                  size_t keysize,
+                                  uint16_t keyid_len,
+                                  uint16_t keysize,
                                   uint8_t *buf,
-                                  size_t buf_len);
-extern size_t asn1_build_prkd_rsa(const uint8_t *label,
-                                  size_t label_len,
+                                  uint16_t buf_len);
+extern uint16_t asn1_build_prkd_rsa(const uint8_t *label,
+                                  uint16_t label_len,
                                   const uint8_t *keyid,
-                                  size_t keyid_len,
-                                  size_t keysize,
+                                  uint16_t keyid_len,
+                                  uint16_t keysize,
                                   uint8_t *buf,
-                                  size_t buf_len);
-extern size_t asn1_build_prkd_aes(const uint8_t *label,
-                                  size_t label_len,
+                                  uint16_t buf_len);
+extern uint16_t asn1_build_prkd_aes(const uint8_t *label,
+                                  uint16_t label_len,
                                   const uint8_t *keyid,
-                                  size_t keyid_len,
-                                  size_t keysize,
+                                  uint16_t keyid_len,
+                                  uint16_t keysize,
                                   uint8_t *buf,
-                                  size_t buf_len);
-extern size_t asn1_build_prkd_generic(const uint8_t *label,
-                                  size_t label_len,
+                                  uint16_t buf_len);
+extern uint16_t asn1_build_prkd_generic(const uint8_t *label,
+                                  uint16_t label_len,
                                   const uint8_t *keyid,
-                                  size_t keyid_len,
-                                  size_t keysize,
+                                  uint16_t keyid_len,
+                                  uint16_t keysize,
                                   int key_tpe,
                                   uint8_t *buf,
-                                  size_t buf_len);
+                                  uint16_t buf_len);
 #endif
